Rework Logger class
