+++ Assets/Scripts/Core/Logger.cs
@@ -1,78 +1,38 @@
-using TMPro;
-using UnityEngine;
+#if DEBUG || UNITY_EDITOR
+#define ENABLE_LOGS
+#endif
+
 using System;
+using System.Diagnostics;
+using UnityEngine;
 
-
-/// <summary>
-/// Write formatted debug data in the Unity console and the Application console
-/// </summary>
-public class Logger : Singleton<Logger>
+namespace ARKitect.Core
 {
-    [SerializeField]
-    private TextMeshProUGUI debugAreaText = null;
-
-    [SerializeField]
-    private bool enableDebug = false;
-
-    [SerializeField]
-    private int maxLines = 15;
-
-    public void Awake()
+    /// <summary>
+    /// Logs a message to the console. 
+    /// This custom class allows to strip the UnityEngine.Debug logging APIs from non-development builds
+    /// and avoid writing to log files.
+    /// See Unity doc: https://docs.unity3d.com/Manual/UnderstandingPerformanceGeneralOptimizations.html
+    /// </summary>
+    public static class Logger
     {
-        if (debugAreaText == null)
+        [Conditional("ENABLE_LOGS")]
+        public static void LogInfo(string message)
         {
-            debugAreaText = GetComponent<TextMeshProUGUI>();
+            UnityEngine.Debug.Log(message);
         }
 
-        debugAreaText.text = string.Empty;
-    }
-
-    public void OnEnable()
-    {
-        debugAreaText.enabled = enableDebug;
-        enabled = enableDebug;
-
-        if (enabled)
+        [Conditional("ENABLE_LOGS")]
+        public static void LogWarning(string message)
         {
-            debugAreaText.text += $"<color=\"white\">{DateTime.Now.ToString("HH:mm:ss.fff")} {this.GetType().Name} enabled.</color>\n";
+            UnityEngine.Debug.Log(message);
+        }
+
+        [Conditional("ENABLE_LOGS")]
+        public static void LogError(string message)
+        {
+            UnityEngine.Debug.Log(message);
         }
     }
 
-    public static void LogInfo(string message)
-    {
-        ClearLines();
-        string line = $"<color=\"green\">{DateTime.Now.ToString("HH:mm:ss.fff")} {message}</color>\n";
-        Instance.debugAreaText.text += line;
-#if UNITY_EDITOR
-        Debug.Log(line);
-#endif
-    }
-
-    public static void LogWarning(string message)
-    {
-        ClearLines();
-        string line = $"<color=\"yellow\">{DateTime.Now.ToString("HH:mm:ss.fff")} {message}</color>\n";
-        Instance.debugAreaText.text += line;
-#if UNITY_EDITOR
-        Debug.LogWarning(line);
-#endif
-    }
-
-    public static void LogError(string message)
-    {
-        ClearLines();
-        string line = $"<color=\"red\">{DateTime.Now.ToString("HH:mm:ss.fff")} {message}</color>\n";
-        Instance.debugAreaText.text += line;
-#if UNITY_EDITOR
-        Debug.LogError(line);
-#endif
-    }
-
-    public static void ClearLines()
-    {
-        if (Instance.debugAreaText.text.Split('\n').Length >= Instance.maxLines)
-        {
-            Instance.debugAreaText.text = string.Empty;
-        }
-    }
 }
