+++ Assets/Scripts/Core/LoggerOld.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1b6f9938e44ef6246a4f013055f5fa79
+guid: 51ad33d464db0d4439b0215f86767ab4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
